backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
xgboost>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
joblib>=1.3.0
